Made the Entity events cancellable

+++ MineLW.Core/Entities/Entity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Numerics;
 using MineLW.API.Entities;
-using MineLW.API.Events.Entities;
+using MineLW.API.Entities.Events;
 using MineLW.API.Worlds.Context;
 
 namespace MineLW.Entities
@@ -15,7 +15,7 @@ namespace MineLW.Entities
         public IWorldContext WorldContext { get; set; }
         public Vector3 Position { get; set; } = Vector3.Zero;
         public Vector2 Rotation { get; set; } = Vector2.Zero;
-        
+
         public event EventHandler<EntityEventArgs> Removed;
         public event EventHandler<EntityWorldEventArgs> WorldChanged;
         public event EventHandler<EntityPositionEventArgs> PositionChanged;
@@ -25,7 +25,7 @@ namespace MineLW.Entities
             Id = id;
             Uuid = uuid;
         }
-        
+
         private void EnsureValid()
         {
             if (!Valid)
@@ -43,8 +43,10 @@ namespace MineLW.Entities
 
             if (position != default && Position != position)
             {
-                PositionChanged?.Invoke(this, new EntityPositionEventArgs(this, Position, position));
-                Position = position;
+                var positionEventArgs = new EntityPositionEventArgs(this, Position, position);
+                PositionChanged?.Invoke(this, positionEventArgs);
+                if (!positionEventArgs.Canceled)
+                    Position = position;
             }
 
             if (rotation != default && Rotation != rotation)
@@ -55,7 +57,11 @@ namespace MineLW.Entities
             if (WorldContext == worldContext)
                 return;
 
-            WorldChanged?.Invoke(this, new EntityWorldEventArgs(this, WorldContext, worldContext));
+            var worldEventArgs = new EntityWorldEventArgs(this, WorldContext, worldContext);
+            WorldChanged?.Invoke(this, worldEventArgs);
+            if (worldEventArgs.Canceled)
+                return;
+
             WorldContext?.EntityManager.RemoveEntity(this);
             WorldContext = worldContext;
             WorldContext.EntityManager.AddEntity(this);
